Add gpiod dependency in poetry
This replaces the system python3-libgpiod dependency, which is not
available in the poetry env by default. The current version of this
package is based on libgpiod 2.x, which has a different API than the
python3-libgpiod package which was still based on 1.x, so this commit
breaks the gpiod backend.