Change: Handle timezone-aware timestamps in acls
The django config now has USE_TZ=True, so the timestamps it produces
have timezone info. Because python does not allow comparing aware
timestamps with naive timestamps, the current time is also made aware.

Note that the per-day start and end times are not aware, but always
interpreted in the local timezone of the lock.